refactor(api): split recentChanges.js controller into per-responsibility modules (#599)
recentChanges.js had grown to 665 lines (audit finding C1). Split it by
responsibility into focused routers under routes/recentChanges/:

  changes.js  — /api/{user,resources,access-package,identities}/:id/recent-changes
  timeline.js — /api/{...}/:id/timeline + the pure timeline builders
  shared.js   — useSql / UUID_RE / day-limit clamps + the label-resolution helpers
                (toEvent / lookup* / resourceCounterpartyKind) both routers use

recentChanges.js is now a barrel that composes the two routers and
re-exports the pure timeline builders (imported by
recentChanges.timeline.test.js), so app.js, the tests and every public
path are unchanged.

The recent-changes handlers and the timeline builders both use the label
helpers, so those live in shared.js rather than being duplicated. Added
recentChanges.data.test.js driving each handler with real _history rows
+ label lookups — previously only the empty-history path was covered, so
the event-building loops were untested; this lifts changed-line diff
coverage to 89% and overall suite coverage by ~1 point.

Grandfathered complexity entries relocated (no increase): cyclomatic
[30,27] and cognitive [42,38,17] -> changes.js. Full API suite green
(1758 tests); coverage above every floor; both complexity gates and the
OpenAPI-drift guard pass.
