+++ .ci/complexity-baseline.json
@@ -54,7 +54,7 @@
     "app/api/src/routes/permissions/grid.js": [
       46
     ],
-    "app/api/src/routes/recentChanges.js": [
+    "app/api/src/routes/recentChanges/changes.js": [
       30,
       27
     ],
