+++ changes/split-fat-route-controllers-599.md
@@ -3,3 +3,4 @@
 - Internal maintainability: split the large admin API controller into focused modules (curated-data export/import, risk-config reads, database maintenance, dashboard stats, and settings) behind a thin barrel. No functional change — every admin endpoint behaves exactly as before.
 - Internal maintainability: split the large identities API controller into focused modules (identity list + columns, identity detail, and per-account analyst overrides) behind a thin barrel. No functional change — every identities endpoint behaves exactly as before.
 - Internal maintainability: split the large tags API controller into focused modules (tag CRUD + assignment, and the users/groups/entity-tags list endpoints) behind a thin barrel. No functional change — every tags/users/groups endpoint behaves exactly as before.
+- Internal maintainability: split the large recent-changes API controller into focused modules (the recent-changes panels and the timeline endpoints, sharing common label-resolution helpers) behind a thin barrel, and added tests covering the event-building paths. No functional change — every recent-changes and timeline endpoint behaves exactly as before.
